+++ CHANGELOG.md
@@ -1,4 +1,20 @@
+## Version 0.0.7
+
+* New calculation of downsample factors
+* Support for 32-bit floating point images
+* 8-bit indexed color images no longer fail, but treated as regular 8-bit (necessary for some .lif files)
+* Revised parsing of z/t values
+* New `dumpMetadata()` method to access OME-XML to help with debugging
+* New `getChannelName(channel)` method
+* Introduced `DummyMetadata()` when creating secondary readers for multithreading
+* Removed autoscaling with use of non-RGB `getBufferedThumbnail`
+* Improved parsing of objective magnification
+* Added new code to test QuPath's use of Bio-Formats
+* Added Bio-Formats version number under 'Help &rarr; Installed extensions'
+* An error is shown when starting QuPath is 'bioformats_package.jar' is missing
+
 ## Version 0.0.6
+
 * Switch to Gradle
 * Added workaround for VSI reading bug that meant channels & z-slices were sometimes confused (available in preferences, turned off by default)
 * Added fix where 4-channel 8-bit images could be converted to RGB, losing the 4th channel information for some commands